refactor(providers): move AWS-specific code into internal/providers/aws/
- Delete credential/aws.go and credential/aws_test.go: ParseRoleARN and
  AWSImpliedDeps were dead code, duplicated by providers/aws/grant.go
  and provider.go respectively
- Remove credential.AWSConfig from types.go: unused legacy type,
  providers/aws has its own Config
- Move run/credential_helper.go to providers/aws/credential_helper.go:
  AWS credential helper script belongs with the provider
- Remove legacy credential.ImpliedDependencies fallback in
  run/manager.go: all providers now implement ImpliedDependencies()
  on the provider interface directly
